backend/requirements.txt
pdfplumber
pandas
dash
dash-bootstrap-components
chromadb
sentence-transformers
pytesseract
docx2txt
openpyxl
plotly
openai
pymupdf
pillow
python-dotenv